==> 04-Data_Exploration.ipynb <==
# Jupyter Notebook Purpose

- With the finalized dataset we have some attributes that need to be "managed"
    - without Latitude and Longitude Information I am unable to link Kaggle data to TTC data - Location data will need to be dropped
    - "Incident" information will need to be re-categorized

## Group 2 Members

- 1. Melissa Hartwick - [Email](mailto:mhartwic@uwaterloo.ca)
- 2. McKinleigh Needham - [Email](mailto:mjneedha@uwaterloo.ca)
- 3. Daniel Adam Cebula  - [Email](mailto:dacebula@uwaterloo.ca)
- 4. Athithian Selvadurai - [Email](mailto:a6selvad@uwaterloo.ca)
- 5. Aravind Kakarala - [Email](mailto:akakaral@uwaterloo.ca)
- 6. Allan Sales - [Email](mailto:asales@uwaterloo.ca)

import pandas as pd
import numpy as np
import os
import requests  # simple HTTP library for Python
import io        # Tool for working with streams (Input/Ouput data)
import matplotlib.pyplot as plt
import glob
import time

%matplotlib inline

# Get FilePaths
cwd = os.getcwd()

Final_Data_Directory = os.path.join(cwd, "FINAL_DATA")

Final_Data = os.path.join(Final_Data_Directory, "2014-2019-TTC-Delay-Data.csv")

# Load the Final Data into DataFrames
df = pd.read_csv(Final_Data, parse_dates=["DateTime", "Date"], low_memory=False)
df.sample(5)

df.info()

# Grouping the "Incident" column

# Remove Location as no Latitude or Longitude data is contained within
df = df.drop(columns=['Location'])

# Remove Route as it is too unique to be used for Machine Learning
df = df.drop(columns=['Route'])

# lets look at incidents
df_incident = df.groupby('Incident').count().sort_values('Identity', ascending=False)
incidents = df['Incident'].unique()
len(incidents)

## Grouping as "Other" the categories with less occurrence
others = []
for i in df_incident['Identity'].index:
    if df_incident['Identity'].loc[i] < 200:
        others.append(i)
len(others)

## Grouping the similar Incidents
t0 = time.time()
df['Incident'] = df.apply(lambda x: 'Other' if str(x['Incident']) in others else x['Incident'], axis=1)
print(f"Done 1 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Other' if str(x['Incident']) == 'Miscellaneous Other' else x['Incident'], axis=1)
print(f"Done 2 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'General Delay' if str(x['Incident']) == 'Miscellaneous General Delays' else x['Incident'], axis=1)
print(f"Done 3 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Speed Control' if 'Speed' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 4 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Late Leaving Garage' if 'Late Leaving Garage' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 5 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Door Problems' if 'Door Problems' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 6 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Door Problems' if 'TR Cab Doors' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 7 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Operator' if 'Operator' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 8 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Injured or ill Customer' if 'Injured or ill Customer' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 9 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Assault' if 'Assault' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 10 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Passenger' if 'Passenger' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 11 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Assault' if 'Robbery' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 12 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Mechanical' if 'Brakes' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 13 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Mechanical' if 'Propulsion System' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 14 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Transportation' if 'Transportation' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 15 - {time.time() - t0}")
df['Incident'] = df.apply(lambda x: 'Track' if 'Track' in str(x['Incident']) else x['Incident'], axis=1)
print(f"Done 16 - {time.time() - t0}")

## Reduction from 156 to 46 categories in 'Incident' column.
df_incident = df.groupby('Incident').count().sort_values('Identity', ascending=False)
len(df_incident)

incidents = df['Incident'].unique()
incidents

df_incident.iloc[:, :1]

# Incidents need to be further dropped down from 45 to 4

- need to further drill it down
    - 1. Mechanical
    - 2. Route Problems
    - 3. Investigation / Emergency
    - 4. General

df.groupby(["Incident"]).count()["Identity"].sort_values(ascending=False).to_frame()

mechanical_list = [
    "Mechanical", "Disorderly Patron", "Door Problems"
    , "Equipment - No Trouble Found", "Unsanitary Vehicle", "Timeout"
    , "Insulated Joint Related Problem", "Body", "Train Control - VOBC"
    , "Air Conditioning", "Rail Cars & Shops Opr. Error"
    , "Escalator/Elevator Incident", "Transportation", "Track"
    , "Station Stairway Incident ", "Signals - Train Stops", "ATC Project"
    , "Signals or Related Components Failure ", "Signal Problem - No Trouble"
    , "Collector Booth Alarm Activated"
]
route_list = [
    "Utilized Off Route", "Late Leaving Garage", "Diversion"
    , "Held By", "Vision", "Transit Control Related Problems"
    , "Storm Trains"
]
invest_emerg_list = [
    "Investigation", "Emergency Services", "Fire/Smoke Plan B - Source TTC"
    , "Emergency Alarm Station Activation"
]
general_list = [
    "General Delay", "Other", "Station Other"
    , "S/E/C Department Other", "Weather Reports / Related Delays", "Training Department Related Delays"
    , "T&S Other", "Ice / Snow Related Problems", "Speed Control", "Operator", "Injured Employee"
    , "Injured or ill Customer", "Passenger", "Disorderly Patron", "Assault"
]

## Grouping the Incidents together
t0 = time.time()
df['Incident'] = df.apply(lambda x: 'Mechanical' if str(x['Incident']) in mechanical_list else x['Incident'], axis=1)
print(f"Done 1 - {time.time() - t0}")

df['Incident'] = df.apply(lambda x: 'Route Problems' if str(x['Incident']) in route_list else x['Incident'], axis=1)
print(f"Done 2 - {time.time() - t0}")

df['Incident'] = df.apply(lambda x: 'Investigation / Emergency' if str(x['Incident']) in invest_emerg_list else x['Incident'], axis=1)
print(f"Done 3 - {time.time() - t0}")

df['Incident'] = df.apply(lambda x: 'General' if str(x['Incident']) in general_list else x['Incident'], axis=1)
print(f"Done 4 - {time.time() - t0}")

df["Incident"] = df["Incident"].fillna("General")
print(f"Done 5 - {time.time() - t0}")

df.groupby(["Incident"]).count()["Identity"].sort_values(ascending=False).to_frame()

# check the distribution across subway, bus and streetcar
df.groupby(["Identity", "Incident"]).count()["DateTime"].to_frame()

# Add the hour in 24 hour format

# Get the nearest hour and add it to the dataset
df["Hour"] = df["DateTime"].dt.round('h').dt.hour

# Reorder the columns and take a subset
df = df.loc[:, ["Identity", "DateTime", "Date", "Hour",
                "BusinessQuarter", "MonthName", "WeekOfYear", "DayName",
                "HolidayName", "Incident", "Temp (°C)",
                "Precip. Amount (mm)", "Wind Dir (10s deg)", "Wind Spd (km/h)", "Stn Press (kPa)",
                "Min Delay", "Min Gap"]]

df.sample(5)

# Write the updated finalized data to a .csv
df.to_csv(os.path.join(Final_Data_Directory, "2014-2019-TTC-Delay-Data-Incident.csv"), index=False)
