==> 05-Preliminary_ML.ipynb <==
# Jupyter Notebook Purpose

- Preliminary ML learning will need to be done
    - a couple of models are attempted and an ensemble is created from the best performing models

## Group 2 Members

- 1. Melissa Hartwick - [Email](mailto:mhartwic@uwaterloo.ca)
- 2. McKinleigh Needham - [Email](mailto:mjneedha@uwaterloo.ca)
- 3. Daniel Adam Cebula  - [Email](mailto:dacebula@uwaterloo.ca)
- 4. Athithian Selvadurai - [Email](mailto:a6selvad@uwaterloo.ca)
- 5. Aravind Kakarala - [Email](mailto:akakaral@uwaterloo.ca)
- 6. Allan Sales - [Email](mailto:asales@uwaterloo.ca)

import pandas as pd
import numpy as np
import os
import requests  # simple HTTP library for Python
import io        # Tool for working with streams (Input/Ouput data)
import matplotlib.pyplot as plt
import glob

%matplotlib inline

# some matplotlib defaults
plt.rcParams['axes.labelsize'] = 14
plt.rcParams['xtick.labelsize'] = 12
plt.rcParams['ytick.labelsize'] = 12

# Get FilePaths
cwd = os.getcwd()

Final_Data_Directory = os.path.join(cwd, "FINAL_DATA")

Final_Data = os.path.join(Final_Data_Directory, "2014-2019-TTC-Delay-Data-Incident.csv")

# Load the Final Data into DataFrames
df = pd.read_csv(Final_Data, parse_dates=["DateTime", "Date"], low_memory=False)
df.sample(5)

df.columns

# only a subset of the columns will be useful for machine learning
df_prelim = df.loc[:,
                   ["Identity", "Hour", "MonthName", "DayName", "HolidayName",
                    "Incident", "Temp (°C)", "Precip. Amount (mm)", "Wind Dir (10s deg)",
                    "Wind Spd (km/h)", "Stn Press (kPa)", "Min Delay", "Min Gap"]]

# subset useful for time series analysis
df_ts = df.loc[:,
               ["DateTime", "Identity", "Incident",
                "Temp (°C)", "Precip. Amount (mm)", "Wind Dir (10s deg)",
                "Wind Spd (km/h)", "Stn Press (kPa)", "Min Delay", "Min Gap"]]

# continuous variables
continuous = ["Temp (°C)", "Precip. Amount (mm)", "Wind Dir (10s deg)",
              "Wind Spd (km/h)", "Stn Press (kPa)"]

# categorical variables
categorical_ml = ["Identity", "Hour", "MonthName", "DayName", "HolidayName", "Incident"]
categorical_ts = ["Identity", "Incident"]

# turn each attribute into categorical
for x in categorical_ml:
    df_prelim[x] = df_prelim[x].astype("category")
    
# turn each attribute into categorical
for x in categorical_ts:
    df_ts[x] = df_ts[x].astype("category")

df_prelim.info()

df_ts.info()

# Preliminary Machine Learning

# Split data into Training and Testing data
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.model_selection import train_test_split

# Scale the data
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import OneHotEncoder

# Imputation of missing values
from sklearn.impute import SimpleImputer

# Scikit pipelines
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer

# Search for best hyperparameters
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV

# Feature Reduction
from sklearn.decomposition import PCA
from sklearn.manifold import TSNE
from sklearn.manifold import LocallyLinearEmbedding
from sklearn.manifold import MDS

# ML Models
from sklearn.neighbors import KNeighborsClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.cluster import KMeans
from sklearn.tree import DecisionTreeClassifier
from sklearn.ensemble import GradientBoostingClassifier

# Train and Test Split

split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=123)
for train_index, test_index in split.split(df_ML, df_ML["DEF_PAY"]):
    strat_train_set = df_ML.loc[train_index]
    strat_test_set = df_ML.loc[test_index]
    
    
# categorical data treatment pipeline
categorical_pipe = Pipeline(steps=[
    ('encoder', OneHotEncoder())
])

# numerical / continuous data treatment pipeline
continuous_pipe = Pipeline(steps=[
    ("scaler", MinMaxScaler())
])

# process the data columns depending on if it is a Numeric or Categorical
preprocess = ColumnTransformer(
   transformers=[
    ('continuous', continuous_pipe, continuous)
   ,('categorical', categorical_pipe, categorical)
])


# assemble it with RandomForestClassifier
pipeline = Pipeline(steps = [
               ('preprocess', preprocess)
              ,('RandomForestClassifier', RandomForestClassifier())
           ])


#Function for X and y choosing the y column (Min Gap or Min Delay)
def x_features (Y, df):
    X = df.loc[:, df.columns != Y]
    y = df[Y].values
    #y = df.loc[:, df.columns == Y]
    if Y == 'Min Gap':
        X = X.drop(['Min Delay'], axis=1)
    else:
        X = X.drop(['Min Gap'], axis=1)
    return X, y

# Running to predict the Min Gap
X, y = x_features('Min Gap', df_cleaned)

from sklearn.model_selection import train_test_split
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)

# Testing Random Forest
from sklearn.ensemble import RandomForestRegressor

rf = RandomForestRegressor(max_depth = 50)
rf.fit(X_train, y_train)

rf.score(X_test, y_test)

# Running to predict the Min Delay
X, y = x_features('Min Delay', df_cleaned)

X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.20)
rf = RandomForestRegressor(max_depth = 15)
rf.fit(X_train, y_train)
rf.score(X_test, y_test)
