==> 07-ML.ipynb <==
# Jupyter Notebook Purpose

- Final ML learning

## Group 2 Members

- 1. Melissa Hartwick - [Email](mailto:mhartwic@uwaterloo.ca)
- 2. McKinleigh Needham - [Email](mailto:mjneedha@uwaterloo.ca)
- 3. Daniel Adam Cebula  - [Email](mailto:dacebula@uwaterloo.ca)
- 4. Athithian Selvadurai - [Email](mailto:a6selvad@uwaterloo.ca)
- 5. Aravind Kakarala - [Email](mailto:akakaral@uwaterloo.ca)
- 6. Allan Sales - [Email](mailto:asales@uwaterloo.ca)

import pandas as pd
import numpy as np
import os
import requests  # simple HTTP library for Python
import io        # Tool for working with streams (Input/Ouput data)
import matplotlib.pyplot as plt
import glob
import time
from joblib import dump, load

%matplotlib inline

# some matplotlib defaults
plt.rcParams['axes.labelsize'] = 14
plt.rcParams['xtick.labelsize'] = 12
plt.rcParams['ytick.labelsize'] = 12

# Get FilePaths
cwd = os.getcwd()

Final_Data_Directory = os.path.join(cwd, "FINAL_DATA")

Final_Data = os.path.join(Final_Data_Directory, "2014-2019-TTC-Delay-Data-Final_ML.csv")

# Load the Final Data into DataFrames
df = pd.read_csv(Final_Data, low_memory=False)
df.sample(5)

# continuous variables
continuous = ["Temp (°C)", "Precip. Amount (mm)"]

# categorical variables
categorical_ml = ["Identity", "Hour", "Season", "Day", "Holiday", "Incident", "Min Delay Category", "Min Gap Category"]

# turn each attribute into categorical
for x in categorical_ml:
    df[x] = df[x].astype("category")

df.info()

# Preliminary Machine Learning Imports

# Split data into Training and Testing data
from sklearn.model_selection import StratifiedShuffleSplit
from sklearn.model_selection import train_test_split

# Scale the data
from sklearn.preprocessing import MinMaxScaler
from sklearn.preprocessing import OneHotEncoder

# Scikit pipelines
from sklearn.pipeline import Pipeline
from sklearn.compose import ColumnTransformer

# Search for best hyperparameters
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection import RandomizedSearchCV

# Feature Reduction
from sklearn.decomposition import PCA

# ML Models
from sklearn.svm import LinearSVC
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier

# Ensemble
from sklearn.ensemble import VotingClassifier

# Classification Report
from sklearn.metrics import classification_report

# Pipeline for both Min Delay and Min Gap

# Stratified Split for Delay Data Categories
split = StratifiedShuffleSplit(n_splits=1, test_size=0.2, random_state=123)

for train_index1, test_index1 in split.split(df, df["Min Delay Category"]):
    delay_train_set = df.loc[train_index1]
    delay_test_set = df.loc[test_index1]
    
# Stratified Split for Delay Gap Categories
for train_index2, test_index2 in split.split(df, df["Min Gap Category"]):
    gap_train_set = df.loc[train_index2]
    gap_test_set = df.loc[test_index2]
    
# Create X (features) and y (labels) for both training and testing data
X_train_delay = delay_train_set.drop(['Min Delay Category', 'Min Gap Category'], axis=1)
X_test_delay = delay_test_set.drop(['Min Delay Category', 'Min Gap Category'], axis=1)
y_train_delay = delay_train_set['Min Delay Category']
y_test_delay = delay_test_set['Min Delay Category']

X_train_gap = gap_train_set.drop(['Min Delay Category', 'Min Gap Category'], axis=1)
X_test_gap = gap_test_set.drop(['Min Delay Category', 'Min Gap Category'], axis=1)
y_train_gap = gap_train_set['Min Gap Category']
y_test_gap = gap_test_set['Min Gap Category']

# one hot encoding categories
one_hot_categories = [
    "Identity", "Hour", "Season", "Day", "Holiday", "Incident"
]

# categorical data treatment pipeline
categorical_pipe = Pipeline(steps=[
    ('encoder', OneHotEncoder())
])

# continuous variables
continuous_columns = [
    'Temp (°C)', 'Precip. Amount (mm)'
]

# numerical / continuous data treatment pipeline
continuous_pipe = Pipeline(steps=[
    ("scaler", MinMaxScaler())
])

# perserve 95% of variance
pca = PCA(n_components=0.95)

preprocess = ColumnTransformer(
    transformers=[
    ('continuous', continuous_pipe, continuous_columns),
    ('categorical', categorical_pipe, one_hot_categories)
    ])

# assemble it with RandomForestClassifier
pipeline_rf_delay = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('RandomForestClassifier', RandomForestClassifier(n_estimators=10))
           ])
pipeline_rf_gap = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('RandomForestClassifier', RandomForestClassifier(n_estimators=10))
           ])

# Assemble it with Gradient Boosted Classifier
pipeline_gbc_delay = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('GradientBoostingClassifier', GradientBoostingClassifier(n_estimators=10))
           ])
pipeline_gbc_gap = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('GradientBoostingClassifier', GradientBoostingClassifier(n_estimators=10))
           ])

# Assemble it with Support Vector Classifier
pipeline_svc_delay  = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('LinearSVC', LinearSVC(tol=0.1))
           ])
pipeline_svc_gap  = Pipeline(steps = [
               ('preprocess', preprocess),
               ("pca", pca),
               ('LinearSVC', LinearSVC(tol=0.1))
           ])

# Create an Ensemble
# Delay
vt_delay = VotingClassifier(
    estimators=[("rf", pipeline_rf_delay),
                ("gbc", pipeline_gbc_delay),
                ("svc", pipeline_svc_delay)],
    voting="hard"
)

# Gap
vt_gap = VotingClassifier(
    estimators=[("rf", pipeline_rf_gap),
                ("gbc", pipeline_gbc_gap),
                ("svc", pipeline_svc_gap)],
    voting="hard"
)

# Train the Model for Delay
t0 = time.time()

vt_delay.fit(X_train_delay, y_train_delay)

print(f"Time: {time.time() - t0}")

# Get Preditions and Classification Report
y_test_delay_pred = vt_delay.predict(X_test_delay)

class_delay = classification_report(y_test_delay, y_test_delay_pred)

print(f"Minutes Delay Classification Report\n\n{class_delay}")

# save the model for later
from joblib import dump, load

dump(vt_delay, 'Minutes_Delay_Model.joblib') 

# Train the Model for Gap
t0 = time.time()

vt_gap.fit(X_train_gap, y_train_gap)

print(f"Time: {time.time() - t0}")

# Get Preditions and Classification Report
y_test_gap_pred = vt_gap.predict(X_test_gap)

class_gap = classification_report(y_test_gap, y_test_gap_pred)

print(f"Minutes Gap Classification Report\n\n{class_gap}")

# save the model for later
from joblib import dump, load

dump(vt_gap, 'Minutes_Gap_Model.joblib')
